pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = "APIRest for trivia management."
authors = ["Eric Durante <eric.alejandro.durante@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.8"
django-oauth-toolkit = "^3.0.1"
celery = {extras = ["rabbitmq"], version = "^5.4.0"}
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
